Ignore the view set by the user by default.
This will make the gui stand still by defauly even when the game view is moving.
When changing resetView to false, you have to make sure that the same view is set when calling handleEvent and when calling draw.
Setting the view on a place where nothing is rendered may not be intuitive and therefore the default view is now used everywhere by default.