The position of the view of the gui is no longer stored in GuiContainer
This change was made because ComboBox would open its list at the wrong location when the center of the view was not the center of the window.

Note that this is a breaking change: the getPosition function of the root container and thus also the getAbsolutePosition of widgets will return a different value than earlier versions when the center of the view was changed.