Rewrote the layout system again
This combines the best features of the last 2 implementations, which were extremely different with the last one focusing on being easy but having dropped several features.

You can still use percentages to refer to the size of the parent:
  widget->setPosition({"20%", "10%"});

But binding widgets has been introduced again so that you can e.g. easily center the widget:
  widget->setPosition("(parent.size - size) / 2");