:sparkles: Moved saving code from WidgetSaver class to the widget itself, which allows the internal state to be saved.
For example, a button previously saved the TextSize which was visible on the screen. If TextSize was 0 internally, it would no longer be 0 after saving and loading it.
The button would still look the same, but when scaling it the text would remain at a fixed size. The auto-sizing text can now be properly saved.

Several small bug fixes got included in this commit, including one breaking change. The getMenus from MenuBar now returns a vector of pairs instead of a map. This was
done to preserve the order of the menus.