Update DataIO.cpp
As it stood, the attribute inheritance feature did not work for arbitrary sections. Even the theme Black.txt used in an example could not be parsed, and the program would exit with an exception once the first colon character was encountered. I suspect this fix to the parseSection function should allow it to recognize the inheritance operator in theme files.