Fixed build failure on macOS with SDL backend
SDL's new cmake config file doesn't provide the SDL2main target in 2.24 on macOS because the main library isn't needed (they will create a dummy target again in the next version).
We now only link to the target when it is found. The target is needed on Windows and iOS (for other platforms it is only a dummy), but we currently don't explicitly verify that the target was indeed found on those platforms.