Changed scaling when object is smaller than corners of 9-slice
Previously the code would fall back to a different scaling type, but this resulted in a completely different looking picture when crossing the minimum size threshold. Now we start gradually shrinking the minimum size picture, so the results just above or just below the threshold still look very similar. Partially related to #257