Avoid return of a function returning void
The function itself returns void as does the function called in the
retun statement, so there's no bug here, just a small passibility of
confusion. When you return something you expect a value to be
returned, which is not the case here, so just call the function and
let that be it.