Use anonymous namespace rather than static
'static' has too many meanings (and was for a while actually proposed to
be deprecated for the meaning of functions with internal linkage;
although that was reversed). In any case, an anonymous namespace
clearly communicates meaning and internal linkage, so let's prefer
that over static functions - the end result is the same, just less
confusing.